+++ .storybook/main.js
@@ -13,6 +13,7 @@ module.exports = {
         '../src/**/**/*.stories.@(js|jsx|ts|tsx)'
     ],
     addons: [
+        'storybook-dark-mode',
         '@storybook/addon-links',
         '@storybook/addon-essentials',
         '@storybook/addon-a11y',