+++ src/components/Menu/hooks/useMenu/useMenu.ts
@@ -1,0 +1,7 @@
+import { useContext } from 'react'
+
+import MenuContext from '../../contexts/MenuContext'
+
+const useMenu = () => useContext(MenuContext)
+
+export default useMenu