+++ src/components/Select/Select.tsx
@@ -41,6 +41,8 @@ const Select: Types.SelectComponent = forwardRef((props, ref) => {
         isMultiSelect = false,
         variant,
         size,
+        disabled,
+        showFilter,
         ...filterdProps
     } = props
 
@@ -50,7 +52,7 @@ const Select: Types.SelectComponent = forwardRef((props, ref) => {
     const [triggerRef, setTriggerRef] = useState<HTMLButtonElement | null>(null)
     const [listRef, setListRef] = useState<HTMLDivElement | null>(null)
     const [triggerWidth, setTriggerWidth] = useState<Types.SelectProps['width']>(width)
-    
+
     const [isOpen, setIsOpen] = useState(!width && !minWidth ? true : false)
 
     useImperativeHandle(ref, () => {
@@ -64,10 +66,10 @@ const Select: Types.SelectComponent = forwardRef((props, ref) => {
         if (!width && !minWidth) {
             if (listRef && triggerWidth === undefined ) {
                 const listRect = listRef.getBoundingClientRect()
-    
+
                 setTriggerWidth(listRect.width)
                 setIsOpen(false)
-                
+
                 shouldfocusTrigger.current = true
             }
         }
@@ -83,7 +85,7 @@ const Select: Types.SelectComponent = forwardRef((props, ref) => {
             if (!isOpen && previousIsOpenState.current && triggerRef) {
                 triggerRef.focus()
             }
-    
+
             previousIsOpenState.current = isOpen
         }
     }, [isOpen, triggerRef])
@@ -128,7 +130,7 @@ const Select: Types.SelectComponent = forwardRef((props, ref) => {
             }
         }
     }, [isOpen, listProps])
-    
+
     const handleOnSelect = useCallback((item: any) => {
         if (typeof onSelect === 'function') {
             onSelect(item)
@@ -157,7 +159,9 @@ const Select: Types.SelectComponent = forwardRef((props, ref) => {
     return (
         <>
             <Styled.Button
+                type={"button"}
                 ref={setTriggerRef}
+                disabled={disabled}
                 {...triggerProps}
                 variant={variant}
                 size={size}
@@ -180,6 +184,7 @@ const Select: Types.SelectComponent = forwardRef((props, ref) => {
                 onClickOutside={handleOnClickOutside}
                 listProps={{
                     ...listProps,
+                    showFilter: showFilter,
                     ref: setListRef,
                     type: ListItemTypes.Select,
                     onSelect: handleOnSelect,
@@ -190,7 +195,7 @@ const Select: Types.SelectComponent = forwardRef((props, ref) => {
                     Children.map(children, (child) => {
                         if (child) {
                             const childElement = child as ReactElement<ListItemProps>
-            
+
                             const childProps = {
                                 ...childElement.props,
                                 selected: selected