+++ src/components/Select/types.ts
@@ -41,7 +41,9 @@ export interface SelectProps extends HTMLAttributes<HTMLDivElement> {
     minWidth?: number | string,
     selected: any[],
     variant?: SelectVariantType,
-    size?: SelectSizeType
+    size?: SelectSizeType,
+    disabled?: boolean,
+    showFilter?: boolean
 }
 
 export type SelectRef = {