+++ src/components/Sheet/style.ts
@@ -1,0 +1,16 @@
+import styled from 'styled-components'
+
+const Wrapper = styled.div`
+    padding: 0;
+    opacity: 0;
+    outline: none;
+    margin: 0 auto;
+    position: absolute;
+    background: #ffffff;
+    overflow: hidden;
+    z-index: 900;
+`
+
+export {
+    Wrapper
+}